+++ dicom_converter/utils/hospital_split.py
